Complete Class Edit And Delete

+++ app/Http/Controllers/ClassController.php
@@ -4,6 +4,10 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\ClassRequest;
 use App\Models\Classes;
+use App\Models\ClassSection;
+use App\Models\ClassGroup;
+use App\Models\Section;
+use App\Models\Group;
 
 class ClassController extends Controller
 {
@@ -32,7 +36,13 @@ class ClassController extends Controller
      */
     public function create()
     {
+        $selected = [ 'id', 'name' ];
+        $sections = Section::active()->get($selected);
+        $groups = Group::active()->get($selected);
+
         $data = [
+            'sections' => $sections,
+            'groups' => $groups,
             'page_title' => 'Create Class',
             'menu' => 'Class'
         ];
@@ -48,7 +58,25 @@ class ClassController extends Controller
      */
     public function store(ClassRequest $request)
     {
-        Classes::create($request->validated()); 
+        // CREATE CLASS
+        $class = Classes::create([ 'name' => $request->name ]);
+        
+        // Save class section
+        collect($request->section_id)->each(function($section_id) use($class) {
+            ClassSection::create([
+                'class_id' => $class->id,
+                'section_id' => $section_id
+            ]);
+        });
+
+        // Svae class groups
+        collect($request->group_id)->each(function($group_id) use($class) {
+            ClassGroup::create([
+                'class_id' => $class->id,
+                'group_id' => $group_id
+            ]);
+        });
+
         return response()->successMessage('Class Created Successfully');
     }
 
@@ -61,9 +89,19 @@ class ClassController extends Controller
     public function edit($id)
     {
         $class = Classes::findOrFail($id);
+        $section_ids = $class->sections->pluck('section_id')->toArray();
+        $group_ids = $class->groups->pluck('group_id')->toArray();
+
+        $selected = [ 'id', 'name' ];
+        $sections = Section::active()->get($selected);
+        $groups = Group::active()->get($selected);
 
         $data = [
             'class' => $class,
+            'sections' => $sections,
+            'groups' => $groups,
+            'section_ids' => $section_ids,
+            'group_ids' => $group_ids,
             'page_title' => 'Edit Class',
             'menu' => 'Class'
         ];
@@ -80,9 +118,35 @@ class ClassController extends Controller
      */
     public function update(ClassRequest $request, $id)
     {
-        $class = Classes::findOrFail($id);
-        $class->update($request->validated());
-        return response()->successMessage('Class Updated Successfully !');
+        $class = Classes::find($id);
+
+        if ($class) {
+            $class->update([ 'name' => $request->name ]);
+
+            // Delete where id not exists in request
+            $class->sections->whereNotIn('section_id', $request->section_id)->each->delete();
+            $class->groups->whereNotIn('group_id', $request->group_id)->each->delete();
+
+            // Save class section
+            collect($request->section_id)->each(function($section_id) use($class) {
+                ClassSection::firstOrCreate([
+                    'class_id' => $class->id,
+                    'section_id' => $section_id
+                ]);
+            });
+
+            // Svae class groups
+            collect($request->group_id)->each(function($group_id) use($class) {
+                ClassGroup::firstOrCreate([
+                    'class_id' => $class->id,
+                    'group_id' => $group_id
+                ]);
+            });
+
+            return response()->successMessage('Class Updated Successfully !');
+        }
+
+        return response()->errorMessage('Class not Found !');
     }
 
     /**
@@ -93,9 +157,14 @@ class ClassController extends Controller
      */
     public function destroy($id)
     {
-        $class = Classes::findOrFail($id);
-        $class->delete();
-        return response()->successMessage('Class Deleted Successfully !');
+        $class = Classes::find($id);
+        
+        if ($class) {
+            $class->delete();
+            return response()->successMessage('Class Deleted Successfully !');
+        }
+
+        return response()->errorMessage('Class not Found !');
     }
 
     /**
